requirements.txt
streamlit
pandas
numpy
altair
python-dateutil
statsmodels
matplotlib